set both Spec and Status fields from attrs
For those APIs like SNS and SQS that use a pattern of storing resource
fields as an Attributes key/value map, we need to set fields on both the
Spec and the Status object from the attributes returned in the
GetAttributes output shape. Otherwise, fields that are "defaulted" on
the server side will not be reflected in the Spec object.